refactor: move CLI specifics into pkg/cli
